pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "automacao-banda-larga"
version = "0.1.0"
description = "Pipeline de prospecção — AutomacaoBandaLarga"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "phonenumbers>=8.13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    # stubs para mypy
    "pandas-stubs>=2.0.0",
    "types-requests>=2.31.0",
]

[tool.setuptools.packages.find]
include = ["utils*", "config*", "domain*", "infra*"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long — já controlado por line-length
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # assert é esperado em testes

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
# Módulos sem stubs — ignora erros de import
[[tool.mypy.overrides]]
module = ["phonenumbers.*", "tqdm.*", "xlsxwriter.*"]
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
